update to use get tenth ms for estimator and change the estimator_module_input_t to use f64

+++ src/application/estimator/estimator_module.h
@@ -27,7 +27,7 @@ typedef struct {
  * input to estimator_module function
  */
 typedef struct {
-	uint32_t timestamp_ms; // new timestamp (ms)
+	float64_t timestamp_ms; // new timestamp (ms)
 	y_imu_t movella; // latest movella data
 	y_imu_t pololu; // latest pololu data
 	bool movella_is_dead; // true if movella is dead
